Move nresty to appropriate location
