+++ service/clients/assistcli/edgebiossystem.go
@@ -3,7 +3,7 @@ package assistcli
 import (
 	"fmt"
 	"github.com/NubeIO/rubix-assist/model"
-	"github.com/NubeIO/rubix-assist/service/clients/assistcli/nresty"
+	"github.com/NubeIO/rubix-assist/service/clients/helpers/nresty"
 )
 
 func (inst *Client) EdgeBiosPing(hostIDName string) (*model.Message, error) {
