+++ service/clients/assistcli/edgesystem.go
@@ -4,7 +4,7 @@ import (
 	"fmt"
 	"github.com/NubeIO/lib-dhcpd/dhcpd"
 	"github.com/NubeIO/lib-networking/networking"
-	"github.com/NubeIO/rubix-assist/service/clients/assistcli/nresty"
+	"github.com/NubeIO/rubix-assist/service/clients/helpers/nresty"
 	"github.com/NubeIO/rubix-edge/service/system"
 )
