+++ service/clients/assistcli/ffstream.go
@@ -3,7 +3,7 @@ package assistcli
 import (
 	"fmt"
 	"github.com/NubeIO/nubeio-rubix-lib-models-go/pkg/v1/model"
-	"github.com/NubeIO/rubix-assist/service/clients/assistcli/nresty"
+	"github.com/NubeIO/rubix-assist/service/clients/helpers/nresty"
 )
 
 // AddStreamToExistingFlow add a stream to an existing flow
