+++ service/clients/edgebioscli/system.go
@@ -3,7 +3,7 @@ package edgebioscli
 import (
 	"fmt"
 	"github.com/NubeIO/rubix-assist/model"
-	"github.com/NubeIO/rubix-assist/service/clients/assistcli/nresty"
+	"github.com/NubeIO/rubix-assist/service/clients/helpers/nresty"
 )
 
 func (inst *BiosClient) Ping() (*model.Message, error) {
