+++ service/clients/edgecli/files.go
@@ -5,7 +5,7 @@ import (
 	"fmt"
 	"github.com/NubeIO/rubix-assist/model"
 	"github.com/NubeIO/rubix-assist/pkg/assistmodel"
-	"github.com/NubeIO/rubix-assist/service/clients/assistcli/nresty"
+	"github.com/NubeIO/rubix-assist/service/clients/helpers/nresty"
 	"os"
 	"strconv"
 )
