+++ service/clients/edgecli/networking.go
@@ -2,7 +2,7 @@ package edgecli
 
 import (
 	"fmt"
-	"github.com/NubeIO/rubix-assist/service/clients/assistcli/nresty"
+	"github.com/NubeIO/rubix-assist/service/clients/helpers/nresty"
 )
 
 func (inst *Client) RestartNetworking() (string, error) {
