+++ service/clients/ffclient/client.go
@@ -3,7 +3,7 @@ package ffclient
 import (
 	"context"
 	"fmt"
-	"github.com/NubeIO/rubix-assist/service/clients/assistcli/nresty"
+	"github.com/NubeIO/rubix-assist/service/clients/helpers/nresty"
 	"github.com/go-resty/resty/v2"
 	"net"
 	"net/http"
