+++ service/clients/ffclient/schema.go
@@ -3,7 +3,7 @@ package ffclient
 import (
 	"encoding/json"
 	"fmt"
-	"github.com/NubeIO/rubix-assist/service/clients/assistcli/nresty"
+	"github.com/NubeIO/rubix-assist/service/clients/helpers/nresty"
 )
 
 // AsJson return as body as blank interface
