+++ service/clients/ffclient/utils.go
@@ -1,7 +1,7 @@
 package ffclient
 
 import (
-	"github.com/NubeIO/rubix-assist/service/clients/assistcli/nresty"
+	"github.com/NubeIO/rubix-assist/service/clients/helpers/nresty"
 )
 
 func (inst *FlowClient) GetQuery(url string) (*[]byte, error) {
